refs(py3): Bump redis and redis-py-cluster to support Python 3.6
Turns out that we only need minor version bumps of both of these to support Python 3.6. The
changelogs don't have anything that looks like a red flag:
 - https://github.com/andymccurdy/redis-py/blob/master/CHANGES#L323
 - https://github.com/Grokzen/redis-py-cluster/blob/master/docs/release-notes.rst#136-nov-16-2018.
   We use 1.3.6 here since it's pinned to 2.10.6 of Redis, which we're bumping to.

There's one long term issue that was preventing us from bumping these, which is covered in
https://github.com/getsentry/sentry/issues/5915. Basically, we pass a `None` value for client when
we create a redis `Script` class, which now breaks since we rely on fetching the encoding from the
client. In most cases I was able to just pass a client through, but in one I've had to write a hack
to work around it, since it's used with RB and isn't easy to pass a client to.